+++ dxr/src/server/handler.rs
@@ -2,6 +2,9 @@ use dxr_shared::{Fault, Value};
 
 use axum::http::HeaderMap;
 
+/// type alias for result type of method handlers
+pub type HandlerResult = Result<Option<Value>, Fault>;
+
 /// trait describing server methods that can be called via XML-RPC
 ///
 /// Handlers for XML-RPC method calls must implement this trait. It is already implemented for `fn`
@@ -12,14 +15,14 @@ use axum::http::HeaderMap;
 pub trait Handler: Send + Sync {
     /// This method is called for handling incoming XML-RPC method requests with the method name
     /// registered for this [`Handler`], with the request's method parameters as its arguments.
-    fn handle(&mut self, params: &[Value], headers: &HeaderMap) -> Result<Value, Fault>;
+    fn handle(&mut self, params: &[Value], headers: &HeaderMap) -> HandlerResult;
 }
 
 /// type alias for plain handler functions without associated data
-pub type HandlerFn = fn(params: &[Value], headers: &HeaderMap) -> Result<Value, Fault>;
+pub type HandlerFn = fn(params: &[Value], headers: &HeaderMap) -> HandlerResult;
 
 impl Handler for HandlerFn {
-    fn handle(&mut self, params: &[Value], headers: &HeaderMap) -> Result<Value, Fault> {
+    fn handle(&mut self, params: &[Value], headers: &HeaderMap) -> HandlerResult {
         self(params, headers)
     }
 }